Add ErrorBoundary to components that can fail and move loading components to ui-kit.tsx
