+++ packages/app/src/App.tsx
@@ -1,4 +1,4 @@
-import { type Component } from "solid-js";
+import { ErrorBoundary, type Component } from "solid-js";
 import { Header } from "~/components/Header";
 import { Editor } from "~/components/Editor";
 import "./index.css";
@@ -7,11 +7,14 @@ import "./index.css";
 import "@fontsource/fira-code";
 import "@fontsource/montserrat";
 import { BreakpointVisualizer } from "./components/BreakpointVisualizer";
+import { LoadingError } from "./components/ui-kit";
 
 const App: Component = () => {
   return (
     <div class="grid h-screen w-screen grid-rows-[auto_1fr]">
-      <Header />
+      <ErrorBoundary fallback={<LoadingError />}>
+        <Header />
+      </ErrorBoundary>
       <Editor />
       {import.meta.env.DEV && <BreakpointVisualizer />}
     </div>
