--- packages/app/src/components/CompilerProvider.tsx
@@ -1,12 +1,0 @@
-// later move more signals to context
-// instead of importing them
-
-//   const compilerOutput = getCompilerOutput();
-
-// export function CompilerProvider(props: FlowProps) {
-//   return (
-//     <CompilerContext.Provider value={compilerOutput}>
-//       {props.children}
-//     </CompilerContext.Provider>
-//   );
-// }
